added broadcasting page
channel can be correctly added to the database. the broadcasting
functionalities are function. but the channel is not removed from the
database after refresh or close the page. due to the pop up is not
working after opening a room. I will rewrite the broadcast service back
to the old structure.